Simplify component events
