+++ library/src/main/java/com/xfinity/blueprint/event/ComponentEvent.kt
@@ -16,7 +16,3 @@ interface ComponentEvent
 interface ComponentEventListener {
     fun onComponentEvent(componentEvent: ComponentEvent) : Boolean
 }
-
-interface ComponentEventEmitter {
-    fun postComponentEvent(componentEvent: ComponentEvent)
-}
